+++ src/const.js
@@ -1,4 +1,4 @@
-export const STORAGES_LIST_ROUTE = '/settings/remote-storage';
+export const STORAGES_LIST_ROUTE = '/settings/remote-storage/configurations';
 
 export const SECTIONS_STORAGE = {
   INFORMATION: 'information',